implementacion de re-jugabilidad

+++ MainMenu.cpp
@@ -5,41 +5,41 @@ void initializeTexts()
 {
 	Texto[0].text = "PLAY";
 	Texto[0].fontSize = 40;
-	Texto[0].pos.x = GetScreenWidth() / 2 - MeasureText(Texto[0].text.c_str(), 50) / 2 + 12;
+	Texto[0].pos.x = GetScreenWidth() / 2 - MeasureText(Texto[0].text.c_str(), 50) / 2.0f + 12;
 	Texto[0].pos.y = 170;
 
-	Texto[0].backgroundBttn.x = GetScreenWidth() / 2 - MeasureText(Texto[0].text.c_str(), 50) / 2;
+	Texto[0].backgroundBttn.x = GetScreenWidth() / 2.0f- MeasureText(Texto[0].text.c_str(), 50) / 2;
 	Texto[0].backgroundBttn.y = 163;
-	Texto[0].backgroundBttn.width = MeasureText(Texto[0].text.c_str(), 50);
+	Texto[0].backgroundBttn.width = static_cast<float>(MeasureText(Texto[0].text.c_str(), 50));
 	Texto[0].backgroundBttn.height = 50;
 
 
 	Texto[1].text = "CREDITS";
 	Texto[1].fontSize = 40;
-	Texto[1].pos.x = GetScreenWidth() / 2 - MeasureText(Texto[1].text.c_str(), 50) / 2 + 22;
+	Texto[1].pos.x = GetScreenWidth() / 2 - MeasureText(Texto[1].text.c_str(), 50) / 2.0f+ 22;
 	Texto[1].pos.y = 310;
 
-	Texto[1].backgroundBttn.x = GetScreenWidth() / 2 - MeasureText(Texto[1].text.c_str(), 50) / 2;
+	Texto[1].backgroundBttn.x = GetScreenWidth() / 2.0f- MeasureText(Texto[1].text.c_str(), 50) / 2;
 	Texto[1].backgroundBttn.y = 303;
-	Texto[1].backgroundBttn.width = MeasureText(Texto[1].text.c_str(), 50);
+	Texto[1].backgroundBttn.width = static_cast<float>(MeasureText(Texto[1].text.c_str(), 50));
 	Texto[1].backgroundBttn.height = 50;
 
 
 	Texto[2].text = "QUIT";
 	Texto[2].fontSize = 40;
-	Texto[2].pos.x = GetScreenWidth() / 2 - MeasureText(Texto[2].text.c_str(), 50) / 2 + 13;
+	Texto[2].pos.x = GetScreenWidth() / 2 - MeasureText(Texto[2].text.c_str(), 50) / 2.0f+ 13;
 	Texto[2].pos.y = 380;
 
-	Texto[2].backgroundBttn.x = GetScreenWidth() / 2 - MeasureText(Texto[2].text.c_str(), 50) / 2;
+	Texto[2].backgroundBttn.x = GetScreenWidth() / 2.0f- MeasureText(Texto[2].text.c_str(), 50) / 2;
 	Texto[2].backgroundBttn.y = 373;
-	Texto[2].backgroundBttn.width = MeasureText(Texto[2].text.c_str(), 50);
+	Texto[2].backgroundBttn.width = static_cast<float>(MeasureText(Texto[2].text.c_str(), 50));
 	Texto[2].backgroundBttn.height = 50;
 }
 
 void drawPlay()
 {
-	DrawRectangle(Texto[0].backgroundBttn.x, Texto[0].backgroundBttn.y, Texto[0].backgroundBttn.width, Texto[0].backgroundBttn.height, WHITE);
-	DrawText(Texto[0].text.c_str(), Texto[0].pos.x, Texto[0].pos.y, Texto[0].fontSize, BLUE);
+	DrawRectangle(static_cast<int>(Texto[0].backgroundBttn.x), static_cast<int>(Texto[0].backgroundBttn.y), static_cast<int>(Texto[0].backgroundBttn.width), static_cast<int>(Texto[0].backgroundBttn.height), WHITE);
+	DrawText(Texto[0].text.c_str(), static_cast<int> (Texto[0].pos.x), static_cast<int>(Texto[0].pos.y), static_cast<int> (Texto[0].fontSize), BLUE);
 }
 
 void checkPlay()
@@ -52,8 +52,8 @@ void checkPlay()
 
 void drawCredits()
 {
-	DrawRectangle(Texto[1].backgroundBttn.x, Texto[1].backgroundBttn.y, Texto[1].backgroundBttn.width, Texto[1].backgroundBttn.height, WHITE);
-	DrawText(Texto[1].text.c_str(), Texto[1].pos.x, Texto[1].pos.y, Texto[1].fontSize, BLUE);
+	DrawRectangle(static_cast<int>(Texto[1].backgroundBttn.x), static_cast<int>(Texto[1].backgroundBttn.y), static_cast<int>(Texto[1].backgroundBttn.width), static_cast<int>(Texto[1].backgroundBttn.height), WHITE);
+	DrawText(Texto[1].text.c_str(), static_cast<int>(Texto[1].pos.x), static_cast<int>(Texto[1].pos.y), static_cast<int>(Texto[1].fontSize), BLUE);
 }
 
 void checkCredits()
@@ -66,8 +66,8 @@ void checkCredits()
 
 void drawQuit()
 {
-	DrawRectangle(Texto[2].backgroundBttn.x, Texto[2].backgroundBttn.y, Texto[2].backgroundBttn.width, Texto[2].backgroundBttn.height, WHITE);
-	DrawText(Texto[2].text.c_str(), Texto[2].pos.x, Texto[2].pos.y, Texto[2].fontSize, RED);
+	DrawRectangle(static_cast<int>(Texto[2].backgroundBttn.x), static_cast<int>(Texto[2].backgroundBttn.y), static_cast<int>(Texto[2].backgroundBttn.width), static_cast<int>(Texto[2].backgroundBttn.height), WHITE);
+	DrawText(Texto[2].text.c_str(), static_cast<int>(Texto[2].pos.x), static_cast<int>(Texto[2].pos.y), static_cast<int>(Texto[2].fontSize), RED);
 }
 
 void checkQuit()
